+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/ViewModels/MainWindowVM.cs
@@ -1,16 +1,22 @@
-﻿using TD.RepositoryContracts;
+﻿using TD.AppInfra.Services;
+using TD.RepositoryContracts;
 using TD.WPF.Commands;
+using TD.WPF.Services;
 using TD.WPF.ViewModels.Trades;
 
 namespace TD.WPF.ViewModels
 {
     internal class MainWindowVM : AbstractDiaryVM
     {
+        private readonly IWindowDialogService _windowDialogService = new WindowDialogService();
+        private ISymbolLookupService _symbolLookupService = new SymbolMockLookupService();
+        private ITradingAccountLookupService _tradingAccountLookupService = new TradingAccountMockLookupService();
+
         public MainWindowVM(ITradeRepository tradeRepository)
         {
             TradeRepository = tradeRepository ?? throw new ArgumentNullException(nameof(tradeRepository));
 
-            TradeHistory = new TradeHistoryVM();
+            TradeHistory = new TradeHistoryVM(tradeRepository, _windowDialogService, _symbolLookupService, _tradingAccountLookupService);
             ShowMaintenanceCommand = new ShowMaintenanceCommand();
             ShowAboutCommand = new ShowAboutCommand();
             ExitCommand = new ExitCommand();