+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/ViewModels/Trades/TradeHistoryVM.cs
@@ -1,6 +1,15 @@
+using TD.AppInfra.Services;
+using TD.RepositoryContracts;
+using TD.WPF.Services;
+
 namespace TD.WPF.ViewModels.Trades
 {
-    internal class TradeHistoryVM : AbstractDiaryVM
+    public class TradeHistoryVM : TradeListVM
     {
+        public TradeHistoryVM(ITradeRepository tradeRepository, IWindowDialogService windowDialogService,
+                              ISymbolLookupService symbolLookupService, ITradingAccountLookupService tradingAccountLookupService)
+            : base(tradeRepository, windowDialogService, symbolLookupService, tradingAccountLookupService)
+        {
+        }
     }
 }